Move exception interceptor to new module
