+++ google/ads/google_ads/client.py
@@ -20,13 +20,11 @@ import grpc
 from collections import namedtuple
 from importlib import import_module
 
-from google.protobuf.message import DecodeError
-
 from google.ads.google_ads import config
 from google.ads.google_ads import oauth2
 
-from google.ads.google_ads.errors import GoogleAdsException
-from google.ads.google_ads.interceptors import MetadataInterceptor
+from google.ads.google_ads.interceptors import MetadataInterceptor, \
+    ExceptionInterceptor
 
 _logger = logging.getLogger(__name__)
 
@@ -228,119 +226,6 @@ class GoogleAdsClient(object):
         service_transport = service_transport_class(channel=channel)
 
         return service_client(transport=service_transport)
-
-
-class ExceptionInterceptor(grpc.UnaryUnaryClientInterceptor):
-    """An interceptor that wraps rpc exceptions."""
-
-    # Codes that are retried upon by google.api_core.
-    _RETRY_STATUS_CODES = (
-        grpc.StatusCode.INTERNAL, grpc.StatusCode.RESOURCE_EXHAUSTED)
-
-    def __init__(self, version=_DEFAULT_VERSION):
-        """Initializes the ExceptionInterceptor
-
-        Args:
-            version: a str of the API version of the request.
-        """
-        self._version = version
-        self._failure_key = (
-            'google.ads.googleads.%s.errors.googleadsfailure-bin' % version)
-
-
-    def _get_google_ads_failure(self, trailing_metadata):
-        """Gets the Google Ads failure details if they exist.
-
-        Args:
-            trailing_metadata: a tuple of metadatum from the service response.
-
-        Returns:
-            A GoogleAdsFailure that describes how a GoogleAds API call failed.
-            Returns None if either the trailing metadata of the request did not
-            return the failure details, or if the GoogleAdsFailure fails to
-            parse.
-        """
-        if trailing_metadata is not None:
-            for kv in trailing_metadata:
-                if kv[0] == self._failure_key:
-                    try:
-                        error_protos = import_module(
-                            'google.ads.google_ads.%s.proto.errors' %
-                                self._version)
-                        ga_failure = error_protos.errors_pb2.GoogleAdsFailure()
-                        ga_failure.ParseFromString(kv[1])
-                        return ga_failure
-                    except DecodeError:
-                        return None
-
-        return None
-
-    def _handle_grpc_failure(self, response):
-        """Attempts to convert failed responses to a GoogleAdsException object.
-
-        Handles failed gRPC responses of by attempting to convert them
-        to a more readable GoogleAdsException. Certain types of exceptions are
-        not converted; if the object's trailing metadata does not indicate that
-        it is a GoogleAdsException, or if it falls under a certain category of
-        status code, (INTERNAL or RESOURCE_EXHAUSTED). See documentation for
-        more information about gRPC status codes:
-        https://github.com/grpc/grpc/blob/master/doc/statuscodes.md
-
-        Args:
-            response: a grpc.Call/grpc.Future instance.
-
-        Raises:
-            GoogleAdsException: If the exception's trailing metadata
-                indicates that it is a GoogleAdsException.
-            RpcError: If the exception's is a gRPC exception but the trailing
-                metadata is empty or is not indicative of a GoogleAdsException,
-                or if the exception has a status code of INTERNAL or
-                RESOURCE_EXHAUSTED.
-            Exception: If not a GoogleAdsException or RpcException the error
-                will be raised as-is.
-        """
-        status_code = response.code()
-        exception = response.exception()
-
-        if status_code not in self._RETRY_STATUS_CODES:
-            trailing_metadata = response.trailing_metadata()
-            google_ads_failure = self._get_google_ads_failure(trailing_metadata)
-
-            if google_ads_failure:
-                request_id = _get_request_id_from_metadata(trailing_metadata)
-
-                raise GoogleAdsException(exception, response,
-                                         google_ads_failure, request_id)
-            else:
-                # Raise the original exception if not a GoogleAdsFailure.
-                raise exception
-        else:
-            # Raise the original exception if error has status code
-            # INTERNAL or RESOURCE_EXHAUSTED.
-            raise exception
-
-    def intercept_unary_unary(self, continuation, client_call_details, request):
-        """Intercepts and wraps exceptions in the rpc response.
-
-        Overrides abstract method defined in grpc.UnaryUnaryClientInterceptor.
-
-        Returns:
-            A grpc.Call instance representing a service response.
-
-        Raises:
-            GoogleAdsException: If the exception's trailing metadata
-                indicates that it is a GoogleAdsException.
-            RpcError: If the exception's trailing metadata is empty or is not
-                indicative of a GoogleAdsException, or if the exception has a
-                status code of INTERNAL or RESOURCE_EXHAUSTED.
-        """
-        response = continuation(client_call_details, request)
-        exception = response.exception()
-
-        if exception:
-            self._handle_grpc_failure(response)
-        else:
-            return response
 
 
 class LoggingInterceptor(grpc.UnaryUnaryClientInterceptor):
